requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
django-environ>=0.11.2
django-allauth>=0.61.1
django-cors-headers>=4.3.1
celery>=5.3.6
redis>=5.0.1
gunicorn>=21.2.0
Pillow>=10.2.0
drf-spectacular>=0.27.1
uvicorn>=0.27.1
django-filter>=24.1
razorpay
